fix #27: moved and consolidated into hooks folder
