+++ src/ts/react/hooks/useTransitionState.ts
