--- src/ts/react/useApproachRate.ts
@@ -1,6 +1,0 @@
-import { loadAr, saveAr, subscribeAr } from "../settings";
-import { useNumericSetting } from "./useSetting";
-
-export function useApproachRate(): [number, (ar: number) => void] {
-  return useNumericSetting(loadAr, saveAr, subscribeAr);
-}
