--- src/ts/react/useHitsoundVolume.ts
@@ -1,6 +1,0 @@
-import { loadHitsoundVolume, saveHitsoundVolume, subscribeHitsoundVolume } from "../settings";
-import { useNumericSetting } from "./useSetting";
-
-export function useHitsoundVolume(): [number, (v: number) => void] {
-  return useNumericSetting(loadHitsoundVolume, saveHitsoundVolume, subscribeHitsoundVolume);
-}
