--- src/ts/react/useVolume.ts
@@ -1,6 +1,0 @@
-import { loadVolume, saveVolume, subscribeVolume } from "../settings";
-import { useNumericSetting } from "./useSetting";
-
-export function useVolume(): [number, (v: number) => void] {
-  return useNumericSetting(loadVolume, saveVolume, subscribeVolume);
-}
